requirements.txt
librosa
lmfit
scipy
sympy
numpy
pandas
matplotlib
playsound
PeakUtils
mpl_pan_zoom
mpl_point_clicker
scikit_learn
scikit_maad
setuptools
ipython
pygobject
ffmpeg